requirements.txt
Package                   Version
------------------------- -----------
aioboto3                  11.3.0
aiobotocore               2.6.0
aiohttp                   3.8.5
aioitertools              0.11.0
aiosignal                 1.3.1
altair                    5.1.1
annotated-types           0.5.0
async-timeout             4.0.3
attrs                     23.1.0
backoff                   2.2.1
beautifulsoup4            4.12.2
blinker                   1.6.2
boto3                     1.28.17
botocore                  1.31.17
Brotli                    1.1.0
cachetools                5.3.1
certifi                   2020.6.20
charset-normalizer        3.2.0
click                     8.1.7
cohere                    4.26.1
contourpy                 1.1.0
cycler                    0.11.0
dataclasses-json          0.5.14
datasets                  2.14.5
deeplake                  3.6.24
dill                      0.3.7
dnspython                 2.4.2
docopt                    0.6.2
entrypoints               0.4
faiss-cpu                 1.7.4
fastavro                  1.8.2
filelock                  3.12.4
fonttools                 4.42.1
frozenlist                1.4.0
fsspec                    2023.6.0
gitdb                     4.0.10
GitPython                 3.1.32
greenlet                  2.0.2
huggingface-hub           0.17.1
humbug                    0.3.2
idna                      3.4
importlib-metadata        6.8.0
Jinja2                    3.1.2
jmespath                  1.0.1
joblib                    1.3.2
Js2Py                     0.74
jsonschema                4.19.0
jsonschema-specifications 2023.7.1
kiwisolver                1.4.5
langchain                 0.0.277
langsmith                 0.0.30
loguru                    0.7.2
markdown-it-py            3.0.0
MarkupSafe                2.1.3
marshmallow               3.20.1
matplotlib                3.7.2
mdurl                     0.1.2
mpmath                    1.3.0
multidict                 6.0.4
multiprocess              0.70.15
mypy-extensions           1.0.0
nest-asyncio              1.5.7
networkx                  3.1
nltk                      3.8.1
numcodecs                 0.11.0
numexpr                   2.8.5
numpy                     1.25.2
openai                    0.27.10
opencv-python             4.8.0.76
packaging                 23.1
pandas                    2.1.0
pathos                    0.3.1
Pillow                    9.5.0
pinecone-client           2.2.4
pip                       23.2.1
pipwin                    0.5.2
pox                       0.3.3
ppft                      1.7.6.7
protobuf                  4.24.2
psutil                    5.9.5
py-cpuinfo                9.0.0
pyarrow                   13.0.0
PyAudio                   0.2.11
pydantic                  2.3.0
pydantic_core             2.6.3
pydeck                    0.8.0
Pygments                  2.16.1
pyjsparser                2.7.1
PyJWT                     2.8.0
Pympler                   1.0.1
pyparsing                 3.0.9
PyPDF2                    3.0.1
PyPrind                   2.11.3
pySmartDL                 1.3.4
PySocks                   1.7.1
python-dateutil           2.8.2
pytz                      2023.3
pytz-deprecation-shim     0.1.0.post0
PyYAML                    6.0.1
referencing               0.30.2
regex                     2023.8.8
requests                  2.31.0
rich                      13.5.2
rpds-py                   0.10.0
s3transfer                0.6.2
scikit-learn              1.3.0
scipy                     1.11.2
seaborn                   0.12.2
setuptools                68.0.0
six                       1.16.0
sklearn                   0.0.post9
smmap                     5.0.0
soupsieve                 2.5
SpeechRecognition         3.10.0
SQLAlchemy                2.0.20
streamlit                 1.26.0
sympy                     1.12
tenacity                  8.2.3
threadpoolctl             3.2.0
tiktoken                  0.4.0
toml                      0.10.2
toolz                     0.12.0
torch                     2.0.1
torchvision               0.15.2
tornado                   6.3.3
tqdm                      4.66.1
typing_extensions         4.7.1
typing-inspect            0.9.0
tzdata                    2023.3
tzlocal                   4.3.1
ultralytics               8.0.180
urllib3                   2.0.4
validators                0.21.2
wheel                     0.38.4
wrapt                     1.15.0
xxhash                    3.3.0
yarl                      1.9.2
zipp                      3.16.2
